+++ src/main/java/roomescape/global/exception/GlobalExceptionHandler.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.payment.exception.InvalidPaymentException;
+import roomescape.payment.global.exception.InvalidPaymentException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
