+++ src/main/java/roomescape/payment/global/domain/PaymentType.java
@@ -1,0 +1,5 @@
+package roomescape.payment.global.domain;
+
+public enum PaymentType {
+    NORMAL, BILLING, BRANDPAY
+}
